Container index mapping within the EMCal Framework
Since many classes store the index of associated or matched objects,
once multiple containers are introduced, there can be an ambiguity
in what index belongs to which container. This can be addressed by
mapping the indices of each object to a unique global index. This commit
introduces the framework for this mapping.

This mapping supports mapping AliEmcalContainer derived classes to
indices or TClonesArrays to indices.

Particle and cluster containers maintain maps for their TClonesArrays to
indices. These can be copied by any task for mapping between their own
AliEmcalContainer derived classes and indices. By copying the maps, it
ensures that the order in which TClonesArrays are added is maintained.

For most users, they will never see mapping, as everything should be
hidden by user facing functions.

See the index mapping class documentation for further information.